+++ types/index.d.ts
@@ -45,6 +45,7 @@ export {
   BlockToolConstructorOptions,
   BlockTool,
   BlockToolData,
+  DefaultBlockToolData,
   Tool,
   ToolConstructable,
   ToolboxConfig,